Finalize events module and update documentation
Remove deprecated code. Extract event actions to undoc'd
actions module. Export `events` from main module.
